Removed the old AtomicXXX functions. This was done as different platforms have different requirements for atomic alignment and such, which is nearly impossible to enforce when we are simply using int* and void* as atomic objects.
The alternative was to create an aligned object could be used in an OOP fashion to perform these actions, however the standard library already does this.

So, instead I removed the AtomicXXX functions and added some typedefs for std::atomic types

The spinlock class was also modified to incorporate these changes.

Hopefully the std::atomic class will allow better platform portability too, as it is implemented properly on all platforms regardless of threadsafe instruction availability